UI in Angular  - Remove the node_Module Folder to upload a single zip file.
Only thing is missing is - it will create a new user but with the existing one. Could have done that part properly but currently directly showing the values in text box by clicking on the edit button , so the selected row edit values will display in text boxes, 
Create New user Save the values based on the Selected row. I wanted do it by uing Reactive forms but could not have much time.